core code updates - use generics collections
git-svn-id: http://svn.code.sf.net/p/hsqldb/svn/base/trunk@6713 7c7dc5f5-a22d-0410-a3af-b41755a11667

+++ src/org/hsqldb/jdbc/JDBCPreparedStatement.java
@@ -443,7 +443,7 @@ implements PreparedStatement {
      */
     public synchronized void setFloat(int parameterIndex,
                                       float x) throws SQLException {
-        setDouble(parameterIndex, (double) x);
+        setDouble(parameterIndex, x);
     }
 
     /**
