+++ src/org/hsqldb/persist/TextFileOps.java
@@ -1,4 +1,4 @@
-/* Copyright (c) 2001-2022, The HSQL Development Group
+/* Copyright (c) 2001-2024, The HSQL Development Group
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -249,6 +249,6 @@ public class TextFileOps {
             } catch (Exception e) {}
         }
 
-        return Result.newUpdateCountResult((int) lineCount);
+        return Result.newUpdateCountResult(lineCount);
     }
 }
